chore(takeover): clean telegram module
