+++ src/service/mod.rs
@@ -38,10 +38,3 @@ pub async fn pumpman(config: &Config, context: Context) -> Result<()> {
 
     Ok(())
 }
-
-/// Disabled the current thread
-pub async fn disable() -> Result<()> {
-    loop {
-        tokio::time::sleep(Duration::from_secs(86400)).await
-    }
-}
