+++ src/telegram/takeover/state.rs
@@ -4,7 +4,7 @@ use crate::{
     model::{Takeover, TakeoverWithCoin},
     telegram::{
         keyboard,
-        takeover::{markup, message, Result, TakeoverDialogue},
+        takeover::{message, Result, TakeoverDialogue},
     },
 };
 use serde::{Deserialize, Serialize};
@@ -35,7 +35,7 @@ pub async fn cto(
     let mint = msg.text().unwrap_or_default().to_string();
     if Pubkey::from_str(&mint).is_err() {
         bot.send_message(msg.chat.id, message::INVALID_ADDRESS)
-            .reply_markup(markup::website("See exists CTOs")?)
+            .reply_markup(message::website("See exists CTOs")?)
             .await?;
 
         return Ok(());
